+++ src/services/permissions/checks/baseCheck/status/BaseCheckStatus.test.ts
@@ -13,14 +13,14 @@ describe('BaseCheck', () => {
   })
 
   it.each(baseCheckScenarios.toTestArray())(
-    `Active caseload: %s | Other caseloads: %s | Roles: %s | Prisoner location: %s | Last Prison location: %s | Date out of last prison: %s | Status: %s`,
+    `Active caseload: %s | Other caseloads: %s | Roles: %s | Prisoner location: %s | Previous Prison location: %s | Date out of previous prison: %s | Status: %s`,
     (
       _activeCaseLoad,
       _otherCaseLoads,
       _roles,
       _prisonerLocation,
-      _lastPrisonId,
-      _dateOutOfLastPrison,
+      _previousPrisonId,
+      _previousPrisonLeavingDate,
       _status,
       testScenario,
     ) => {