+++ CHANGELOG.md
@@ -4,7 +4,11 @@ Please use this to capture reasoning behind changes:
 
 ## 1.0.0
 
-This release changes the permissions around global access to case notes. Previously if a user had both the `POM` and `GLOBAL_SEARCH` roles they would be able read/write case notes any prisoner. The permissions have now been updated so that users with both the `POM` and `GLOBAL_SEARCH` roles will only be able to read/write a prisoner's case notes if the prisoner has been in the users establishment within the last 30 days.
+This release changes the permissions around global access to case notes and read access to religion data.
+
+**Case Notes Permissions:**
+
+Previously if a user had both the `POM` and `GLOBAL_SEARCH` roles they would be able read/write case notes any prisoner. The case notes permissions have now been updated so that users with both the `POM` and `GLOBAL_SEARCH` roles will only be able to read/write a prisoner's case notes if the prisoner has been in the users establishment within the last 30 days.
 
 The permissions check requires additional data to be passed in via the `Prisoner` which now expects the `previousPrisonId` and `previousPrisonLeavingDate` fields to be present. This data is being added to the prisoner search so clients using the permissions library will need to ensure the following fields are present.
 
@@ -18,6 +22,10 @@ export default interface Prisoner {
   previousPrisonLeavingDate?: string
 }
 ```
+**Religion and Belief Permissions:**
+
+The permissions check for `PersonProtectedCharacteristicsPermission.read_religion_and_belief` updated to only allow read access to a
+prisoner's religion data to users who are part of the prisoner's caseload.
 
 ## 0.2.0