+++ CHANGELOG.md
@@ -2,10 +2,11 @@
 
 Please use this to capture reasoning behind changes:
 
-## 1.0.0
+## 1.0.0-alpha.1
 
-Permissions check for `PersonProtectedCharacteristicsPermission.read_religion_and_belief` updated to only allow read access to a
-prisoner's religion data to users who are part of the prisoner's caseload.
+Alpha release to test updated religion and belief permissions.
+
+The permissions check for `PersonProtectedCharacteristicsPermission.read_religion_and_belief` has been updated to only allow read access to a prisoner's religion data to users who are part of the prisoner's caseload.
 
 ## 0.2.0